+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Patch Changes
 
+- ai@4.0.3
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@4.0.2
 
 ## 0.0.1